come back to what used to worked with seuillage

+++ biapy/engine/metrics.py
@@ -616,8 +616,10 @@ class SoftclDiceLoss(nn.Module):
     def forward(self, inputs: torch.Tensor, targets: torch.Tensor) -> torch.Tensor:
         # Appliquer la sigmoïde comme dans DiceLoss
         inputs = F.sigmoid(inputs) 
-        inputs = (inputs > 0.5)
-        inputs = inputs.astype(torch.uint8) #masque
+        inputs = (inputs > 0.5).float()
+        inputs = inputs.to(torch.uint8)
+        # inputs = (inputs > 0.5)
+        # inputs = inputs.astype(torch.uint8) #masque
         #print(np.unique(inputs))
         #inputs = inputs.to(torch.uint8) #la on doit avoir des masque 1 ou 0
